Shape: Add circumradius getter
The circumradius (defined as the radius of the circumscribed circle) is
a useful definition for both rectangles and circles (the two currently
implemented Shape classes).

SPOT/SONIA often needs the diameter of an obstacle shape, which is currently
done by dynamically casting the obstacle shape to a Rectangle and the calculating
the diagonal length like this:

double rectLen = dynamic_cast<Rectangle&>(obstacle->getGeoShape()).getLength();
double rectWidth = dynamic_cast<Rectangle&>(obstacle->getGeoShape()).getWidth();
double diagonal = std::hypot(rectLen, rectWidth);

Instead of directly adding a shape-specific getDiagonal() method to the Rectangle
class, the circumradius should be a more useful addition since it does not
depend on a specific shape.