Add Git pre-commit hook for automatic clang-format checks
A pre-commit hook is a script that is run every time before a commit.
This commit adds a pre-commit hook that checks whether the clang-format
check (as run in the CI) reports any formatting violations, and offers
to fix them automatically. This should help reduce the number of reformatting
commits due to clang-format CI failures.

The hook needs to be installed manually, since the contents of the .git
directory (where the hook needs to be installed) are not part of version control.
The helper script setup_git_hooks.sh automates the installation.