findClosestIndex: Improve distance calculation
The previous algorithm in findClosestIndex used linear space (in the number of
center vertices), but it is possible to implement it using constant space.
Furthermore, the square root calculation can be skipped when calculating distances
only for comparison.
Since findClosestIndex can be called quite often depending on the use case,
this optimization can have a notable impact.